# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
    AuditEntityType:
      type: string
      enum:
      - EPIC
      - FEATURE
      - TASK
    AuditEntryDto:
      type: object
      properties:
        id:
          type: string
        entityType:
          $ref: "#/components/schemas/AuditEntityType"
        entityId:
          type: string
        epicId:
          type: string
        operation:
          $ref: "#/components/schemas/AuditOperation"
        changedBy:
          type: string
        changedAt:
          $ref: "#/components/schemas/Instant"
        snapshot:
          type: string
    AuditOperation:
      type: string
      enum:
      - CREATE
      - UPDATE
      - DELETE
...
    CreateEpicRequest:
      type: object
      required:
      - title
      properties:
        title:
          type: string
          pattern: \S
        description:
          type: string
    CreateFeatureRequest:
      type: object
      required:
      - title
      properties:
        title:
          type: string
          pattern: \S
        description:
          type: string
        dependsOnFeatureId:
          type: string
...
    CreateTaskRequest:
      type: object
      required:
      - repositoryId
      - title
      properties:
        repositoryId:
          type: string
          pattern: \S
        title:
          type: string
          pattern: \S
        description:
          type: string
        dependsOnTaskId:
          type: string
...
    Entry:
      type: object
      properties:
        feature:
          $ref: "#/components/schemas/FeatureDto"
    Entry1:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/TaskDto"
    Entry2:
      type: object
      properties:
        epic:
          $ref: "#/components/schemas/EpicDto"
...
    EpicDto:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        slug:
          type: string
        status:
          type: string
        supersededByEpicId:
          type: string
        description:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
    FeatureDto:
      type: object
      properties:
        id:
          type: string
        epicId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        dependsOnFeatureId:
          type: string
        implementedOn:
          $ref: "#/components/schemas/Instant"
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    Response:
      type: object
      properties:
        epic:
          $ref: "#/components/schemas/EpicDto"
    Response1:
      type: object
      properties:
        epic:
          $ref: "#/components/schemas/EpicDto"
        successor:
          $ref: "#/components/schemas/EpicDto"
    Response10:
      type: object
      properties:
        success:
          type: boolean
    Response11:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/TaskDto"
    Response12:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry2"
    Response13:
      type: object
      properties:
        epic:
          $ref: "#/components/schemas/EpicDto"
...
    Response2:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
...
    Response3:
      type: object
      properties:
        epic:
          $ref: "#/components/schemas/EpicDto"
...
    Response4:
      type: object
      properties:
        success:
          type: boolean
    Response5:
      type: object
      properties:
        feature:
          $ref: "#/components/schemas/FeatureDto"
    Response6:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AuditEntryDto"
    Response7:
      type: object
      properties:
        feature:
          $ref: "#/components/schemas/FeatureDto"
    Response8:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/Entry1"
    Response9:
      type: object
      properties:
        feature:
          $ref: "#/components/schemas/FeatureDto"
...
    TaskDto:
      type: object
      properties:
        id:
          type: string
        featureId:
          type: string
        repositoryId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        dependsOnTaskId:
          type: string
        implementedAt:
          $ref: "#/components/schemas/Instant"
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
    TransitionEpicRequest:
      type: object
      properties:
        target:
          type: string
    UpdateEpicRequest:
      type: object
      required:
      - title
      properties:
        title:
          type: string
          pattern: \S
        description:
          type: string
    UpdateFeatureRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dependsOnFeatureId:
          type: string
        clearDependsOn:
          type: boolean
        implementedOn:
          $ref: "#/components/schemas/Instant"
        clearImplementedOn:
          type: boolean
...
paths:
  /projects/api/epics/{epicId}/features:
    get:
      parameters:
      - name: epicId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response2"
      summary: List Features
      tags:
      - Epic Controller
    post:
      parameters:
      - name: epicId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeatureRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response5"
        "400":
          description: Bad Request
      summary: Create Feature
      tags:
      - Epic Controller
  /projects/api/epics/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEpicRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Epic Controller
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response3"
      summary: Get
      tags:
      - Epic Controller
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response4"
      summary: Delete
      tags:
      - Epic Controller
  /projects/api/epics/{id}/audit:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response6"
      summary: Audit
      tags:
      - Epic Controller
  /projects/api/epics/{id}/transition:
    post:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionEpicRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response1"
        "400":
          description: Bad Request
      summary: Transition
      tags:
      - Epic Controller
...
  /projects/api/features/{featureId}/tasks:
    get:
      parameters:
      - name: featureId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response8"
      summary: List Tasks
      tags:
      - Feature Controller
    post:
      parameters:
      - name: featureId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response11"
        "400":
          description: Bad Request
      summary: Create Task
      tags:
      - Feature Controller
  /projects/api/features/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeatureRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response7"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Feature Controller
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response9"
      summary: Get
      tags:
      - Feature Controller
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response10"
      summary: Delete
      tags:
      - Feature Controller
...
  /projects/api/projects/{projectId}/epics:
    get:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response12"
      summary: List
      tags:
      - Project Epics Controller
    post:
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEpicRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response13"
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Project Epics Controller
...
info:
  title: qits-projects API
  version: 2026.808.110015